# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aleph · P.R.O.A.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/cases/{case_id}": {
      "get": {
        "tags": [
          "casos"
        ],
        "summary": "Ver un caso (queda auditado)",
        "operationId": "get_case_api_cases__case_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "401": {
            "description": "Falta el token o no es válido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "El rol no permite la acción",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "No existe (o no pertenece al caso)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "casos"
        ],
        "summary": "Editar un caso abierto",
        "operationId": "update_case_api_cases__case_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            }
          },
          "401": {
            "description": "Falta el token o no es válido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "El rol no permite la acción",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "No existe (o no pertenece al caso)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "casos"
        ],
        "summary": "Borrar un caso vacío (admin)",
        "operationId": "delete_case_api_cases__case_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Falta el token o no es válido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "El rol no permite la acción",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "No existe (o no pertenece al caso)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cases/{case_id}/captures": {
      "post": {
        "tags": [
          "lens"
        ],
        "summary": "Captura en vivo desde la extensión",
        "description": "Idempotente: la extensión puede reenviar publicaciones ya vistas. Las `interactions` quedan\ncomo relaciones entre las entidades de cuenta. `text_snippets` se conserva en el crudo de la fuente.",
        "operationId": "capture_api_cases__case_id__captures_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureBatch"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureSummary"
                }
              }
            }
          },
          "401": {
            "description": "Falta el token o no es válido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "El rol no permite la acción",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "No existe (o no pertenece al caso)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/captures/lookup": {
      "post": {
        "tags": [
          "lens"
        ],
        "summary": "Qué sabe el caso de los handles visibles en pantalla",
        "operationId": "lookup_api_cases__case_id__captures_lookup_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandleLookupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandleLookupResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta el token o no es válido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "El rol no permite la acción",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "No existe (o no pertenece al caso)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{case_id}/sections": {
      "get": {
        "tags": [
          "lens"
        ],
        "summary": "Incisos del expediente, con contador de hallazgos",
        "operationId": "list_sections_api_cases__case_id__sections_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectionOut"
                  },
                  "title": "Response List Sections Api Cases  Case Id  Sections Get"
                }
              }
            }
          },
          "401": {
            "description": "Falta el token o no es válido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "El rol no permite la acción",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "No existe (o no pertenece al caso)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "lens"
        ],
        "summary": "Crear inciso",
        "operationId": "create_section_api_cases__case_id__sections_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SectionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectionOut"
                }
              }
            }
          },
          "401": {
            "description": "Falta el token o no es válido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "El rol no permite la acción",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "No existe (o no pertenece al caso)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaptureBatch": {
        "properties": {
          "page_url": {
            "type": "string",
            "title": "Page Url"
          },
          "page_title": {
            "type": "string",
            "title": "Page Title",
            "default": ""
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          },
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/AccountProfile"
            },
            "type": "array",
            "title": "Profiles"
          },
          "interactions": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array",
              "maxItems": 3,
              "minItems": 3
            },
            "type": "array",
            "title": "Interactions"
          },
          "text_snippets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Text Snippets"
          },
          "extension_version": {
            "type": "string",
            "title": "Extension Version",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "page_url",
          "platform",
          "captured_at"
        ],
        "title": "CaptureBatch",
        "description": "Lo que la extensión envía mientras el analista navega: POST /api/cases/{id}/captures"
      },
...
      "CaseDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "legal_basis": {
            "type": "string",
            "title": "Legal Basis"
          },
          "tlp": {
            "type": "string",
            "title": "Tlp"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "legal_basis",
          "tlp",
          "status",
          "created_by",
          "created_at"
        ],
        "title": "CaseDetail"
      },
...
      "CaseOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "legal_basis": {
            "type": "string",
            "title": "Legal Basis"
          },
          "tlp": {
            "type": "string",
            "title": "Tlp"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "legal_basis",
          "tlp",
          "status",
          "created_by",
          "created_at"
        ],
        "title": "CaseOut"
      },
      "CaseUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "legal_basis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Basis"
          },
          "tlp": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "clear",
                  "green",
                  "amber",
                  "amber+strict",
                  "red"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Tlp"
          }
        },
        "type": "object",
        "title": "CaseUpdate"
      },
...
      "HandleLookupRequest": {
        "properties": {
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "handles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Handles"
          }
        },
        "type": "object",
        "required": [
          "platform",
          "handles"
        ],
        "title": "HandleLookupRequest",
        "description": "POST /api/cases/{id}/captures/lookup: qué sabe el caso de los handles visibles."
      },
...
      "Message": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "Message"
      },
...
      "ValidationErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ValidationErrorOut"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Token devuelto por POST /api/auth/login",
        "scheme": "bearer"
      }
    }
  }
}